feat: add layoutCfg refactor: extract dagre methods
