+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@zazuko/vue-graph-layout",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "author": "Zazuko (https://zazuko.com)",
   "main": "src/index.js",
   "files": [
